Rework dashboard to handle org flow
