+++ apps/dashboard/src/lib/supabase/supabase-types.ts
@@ -2,3 +2,4 @@ import type { Database } from '$lib/types/database.types';
 
 export type URLInfo = Database['public']['Tables']['urls']['Row'];
 export type LinkInfo = Database['public']['Tables']['dynamic_links']['Row'];
+export type OrgInfo = Database['public']['Tables']['organizations']['Row'];
