+++ apps/dashboard/src/routes/projects/[id]/+page.ts
@@ -1,0 +1,12 @@
+import { redirect } from '@sveltejs/kit';
+import type { PageLoad } from './$types';
+
+export const load: PageLoad = async ({ parent }) => {
+	const { session, supabase } = await parent();
+
+	if (!session) {
+		throw redirect(303, '/login');
+	}
+
+	return { supabase, session };
+};
