+++ apps/dashboard/src/routes/projects/[id]/modals/LinkDetailModal.svelte
